third_party/WebKit/LayoutTests/http/tests/devtools/elements/styles-2/inherited-mixed-case-properties-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that non-standard mixed-cased properties are displayed in the Styles pane.

color: rgb(0, 0, 0);
display: block;
    block - div user agent stylesheet
-webkit-font-smoothing: antialiased;
[expanded] 
element.style { ()

[expanded] 
div { (user agent stylesheet)
    display: block;

======== Inherited from div#container ========
[expanded] 
Style Attribute { ()
    CoLoR: blAck;

[expanded] 
#container { (<style>…</style>)
    -webkit-FONT-smoothing: antialiased;
